Extract DatabaseUtils
